Restructured package and directory structure
Updated Gradle to 7.2.1
Optimized imports
