+++ android/Drawer/app/build.gradle
@@ -6,7 +6,7 @@ android {
     compileSdk 32
 
     defaultConfig {
-        applicationId "com.example.drawer"
+        applicationId "com.drawer"
         minSdk 29
         targetSdk 32
         versionCode 1
@@ -27,7 +27,7 @@ android {
         targetCompatibility JavaVersion.VERSION_1_8
     }
 
-    namespace 'com.example.drawer'
+    namespace 'com.drawer'
 }
 
 dependencies {
