+++ android/Drawer/app/src/androidTest/java/com/drawer/ManualControlInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.drawer;
+package com.drawer;
 
 import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.action.ViewActions.click;
@@ -14,6 +14,9 @@ import static androidx.test.espresso.matcher.ViewMatchers.withText;
 import androidx.test.espresso.intent.Intents;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
+import com.drawer.activities.DrawControl;
+import com.drawer.activities.MainActivity;
+import com.drawer.activities.ManualControl;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
