+++ android/Drawer/app/src/main/java/com/drawer/canvas/CanvasGrid.java
@@ -1,4 +1,4 @@
-package com.example.drawer;
+package com.drawer.canvas;
 
 import android.content.Context;
 import android.graphics.Canvas;
@@ -9,6 +9,7 @@ import android.util.Log;
 import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
+import com.drawer.connectivity.MQTTController;
 import java.util.LinkedList;
 import java.util.Queue;
