+++ android/Drawer/app/src/main/java/com/drawer/runnables/ManualRecordingRun.java
@@ -1,7 +1,8 @@
-package com.example.drawer;
+package com.drawer.runnables;
 
 import android.os.SystemClock;
 import android.widget.Chronometer;
+import com.drawer.connectivity.MQTTController;
 import java.util.ArrayList;
 
 public class ManualRecordingRun implements Runnable {
