+++ android/Drawer/app/src/test/java/com/drawer/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.drawer;
+package com.drawer;
 
 import static org.junit.Assert.assertEquals;
